+++ MedianFilterParallel/binning.h
@@ -9,14 +9,14 @@ namespace hpcparallel
 
 	private:
 		int* bins;
-		size_t resolution;
+		int resolution;
 		hpcparallel::inputstream inputstream;
 		int numThreads;
 		int chunkSize;// buffer size
 
 
 	public:
-		binning(size_t res, char* filename);
+		binning(int res, char* filename);
 
 		hpc::array<int>* processBin();
