+++ MedianFilterParallel/printcsv.h
@@ -4,12 +4,12 @@
 
 namespace hpc
 {
-	void printcsv(size_t resolution, int* bins);
+	void printcsv(int resolution, int* bins);
 
-	void printFileCsv(size_t resolution, int* bins, const char* filename);
+	void printFileCsv(int resolution, int* bins, const char* filename);
 
 	const std::string currentDateTime();
 
-	void writeResultsFile(char* filename, size_t binres, size_t filtersize, unsigned int binningTime, unsigned int smoothingTime);
+	void writeResultsFile(char* filename, int binres, int filtersize, unsigned int binningTime, unsigned int smoothingTime);
 }
