+++ MedianFilterParallel/structs.h
@@ -6,10 +6,10 @@ namespace hpc
 	template < typename T >
 	struct array
 	{
-		size_t size;
+		int size;
 		T* pointer;
 
-		array(size_t s, T* p) : size(s), pointer(p) {}
+		array(int s, T* p) : size(s), pointer(p) {}
 		array() {}
 	};
 }
